requirements.txt
torch
torchvision
fiona
rasterio
geopandas
shapely
numpy
opencv-python
scipy
tqdm
scikit-image
Pillow
